moves complex logic from api.query to fetch.fetchQuery in vizbuilder
